fixed copyright, other comments

+++ tests/models/nn/test_item_net.py
@@ -1,3 +1,17 @@
+#  Copyright 2024 MTS (Mobile Telesystems)
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+
 import typing as tp
 
 import numpy as np
@@ -103,13 +117,14 @@ class TestCatFeaturesItemNet:
         actual_feature_dense = cat_item_embeddings.get_dense_item_features(items)
         expected_feature_dense = torch.tensor(
             [
-                [1.0, 0.0, 1.0, 0.0, 0.0],
-                [1.0, 0.0, 0.0, 1.0, 0.0],
-                [0.0, 1.0, 1.0, 0.0, 0.0],
-                [1.0, 0.0, 0.0, 0.0, 1.0],
-                [0.0, 1.0, 0.0, 1.0, 0.0],
-                [0.0, 1.0, 0.0, 0.0, 1.0],
-            ]
+                [1, 0, 1, 0, 0],
+                [1, 0, 0, 1, 0],
+                [0, 1, 1, 0, 0],
+                [1, 0, 0, 0, 1],
+                [0, 1, 0, 1, 0],
+                [0, 1, 0, 0, 1],
+            ],
+            dtype=torch.float,
         )
 
         assert torch.equal(actual_feature_dense, expected_feature_dense)
